Integer percents add up to 100, no animation

+++ app/src/main/java/com/teplyakova/april/telegramcontest/Drawing/Area.java
@@ -5,7 +5,7 @@ import com.teplyakova.april.telegramcontest.Data.LineData;
 
 public class Area {
 	public LineData Line;
-	public float[] Percentages;
+	public int[] Percentages;
 	public float PosYCoefficient;
 	public float[] MappedPointsY;
 	public android.graphics.Path Path;
